# api-docs.json
{"openapi":"3.1.0","info":{"title":"Delivery Service API","description":"API for managing deliveries by stores and couriers","version":"0.2"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Courier","description":"Courier operations for managing deliveries"},{"name":"Store","description":"Store operations for managing orders"},{"name":"Authentication","description":"Authentication endpoints for stores and couriers"},{"name":"Customer","description":"Customer operations for creating orders"}],"paths":{"/api/store/orders/{orderId}/collect":{"post":{"tags":["Store"],"summary":"Mark order as collected","operationId":"collectOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/store/orders/{orderId}/cancel":{"post":{"tags":["Store"],"summary":"Cancel an order","description":"Cancel a specific order that hasn't been collected yet","operationId":"cancelOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/customer/orders":{"post":{"tags":["Customer"],"summary":"Create a new order","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/api/courier/orders/{orderId}/deliver":{"post":{"tags":["Courier"],"summary":"Mark order as delivered","description":"Complete the delivery of a specific order","operationId":"deliverOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/auth/store/register":{"post":{"tags":["Authentication"],"summary":"Register a new store","operationId":"registerStore","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/auth/store/login":{"post":{"tags":["Authentication"],"summary":"Login as a store","operationId":"loginStore","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/courier/register":{"post":{"tags":["Authentication"],"summary":"Register a new courier","operationId":"registerCourier","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/auth/courier/login":{"post":{"tags":["Authentication"],"summary":"Login as a courier","operationId":"loginCourier","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/store/profile":{"get":{"tags":["Store"],"summary":"Get store's profile","description":"Retrieve the authenticated store's profile information","operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StoreResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/store/orders":{"get":{"tags":["Store"],"summary":"Get store's orders","description":"Retrieve orders associated with the authenticated store","operationId":"getOrders","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResponseOrder"}}}}},"security":[{"bearerAuth":[]}]}},"/api/store/orders/{orderId}":{"get":{"tags":["Store"],"summary":"Get store's order by ID","description":"Retrieve a specific order associated with the store by its ID","operationId":"getOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/courier/ready":{"get":{"tags":["Courier"],"summary":"Set courier status to ready","description":"Update the authenticated courier's status to ready for accepting deliveries","operationId":"makeCourierReady","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourierResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/courier/profile":{"get":{"tags":["Courier"],"summary":"Get courier's profile","description":"Retrieve the authenticated courier's profile information","operationId":"getProfile_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourierResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/courier/orders":{"get":{"tags":["Courier"],"summary":"Get courier's orders","description":"Retrieve orders assigned to the authenticated courier","operationId":"getOrders_1","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResponseOrder"}}}}},"security":[{"bearerAuth":[]}]}},"/api/courier/orders/{orderId}":{"get":{"tags":["Courier"],"summary":"Get courier's order by ID","description":"Retrieve a specific order assigned to the authenticated courier by its ID","operationId":"getOrder_1","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"OrderResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customerEmail":{"type":"string"},"orderItemIds":{"type":"array","items":{"type":"integer","format":"int64"}},"status":{"type":"string","enum":["CREATED","COLLECTED","IN_DELIVERY","DELIVERED","CANCELLED","REFUNDED"]},"createdAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"},"storeId":{"type":"integer","format":"int64"},"courierId":{"type":"integer","format":"int64"}}},"CreateOrderRequest":{"type":"object","properties":{"customerEmail":{"type":"string"},"orderItemIds":{"type":"array","items":{"type":"integer","format":"int64"}},"storeId":{"type":"integer","format":"int64"}},"required":["customerEmail","orderItemIds","storeId"]},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":2},"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","name","password"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"StoreResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"name":{"type":"string"}}},"EntityApiResponseOrder":{"type":"object"},"PagedResponseOrder":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityApiResponseOrder"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"pageNumberOfElements":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"}}},"CourierResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"name":{"type":"string"},"status":{"type":"string","enum":["READY","BUSY","NOT_READY"]},"lastAssignment":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}